create robot name
"""
from Robot init removed robot paths, add name
add method as_dict for robot and state(WIP)
add function get_robot_names
get_robots_paths moved to frontend
"""
